Upload huge application overhaul | Directory nearing completion |
Yeah this code was written over the last few days and because I was testing a lot of things to conform to the style I wanted for the application I decided to not push it to github. The only work left to do on directory is add a dynamic search function that sets visibility states of components based upon the search result.